requirements.txt
pandas>=2.0.0

# 구글 드라이브 API 관련 패키지
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.115.0
